prepare/kg_dataloader/requirements.txt
python-dotenv==1.0.0
neo4j==5.15.0 